pyproject.toml
[project]
name = "skills-hub-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.3",
    "pytest>=8.0",
    "pytest-xdist>=3.5",
    "openai>=1.0",
]
